\docType{package}
\name{amritaPackage-package}
\alias{amritaPackage}
\alias{amritaPackage-package}
\title{The amritaPackage Package}
\description{
  Quasi-bootstrap association tests.
}
\details{
  \tabular{ll}{ Package: \tab amritaPackage\cr Type: \tab
  Package\cr Version: \tab 0.01-1\cr Date: \tab
  2013-10-09\cr Depends: \tab R (>= 2.10.0), kinship2,
  matrixcalc\cr Encoding: \tab UTF-8\cr License: \tab GPL
  (>= 3)\cr LazyLoad: \tab yes\cr URL: \tab
  http://stanford.edu/~amray/thepackage/index.html\cr }

  Provides quasi-bootstrap p-values for provided and
  user-provided tests.
}
\author{
  Amrita Ray \email{amray@stanford.edu}
}
\keyword{package}

